# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PalPanel API
  version: 1.0.0
  description: Stable HTTP contract for the PalPanel 1.x management API.
servers:
  - url: /api
security:
  - sessionCookie: []
  - developmentKey: []
...
paths:
...
  /server/save:
    post: {operationId: saveWorld, tags: [server], x-palpanel-permission: 'server:control', responses: *writeResponses}
...
  /save-sources:
    get: {operationId: listSaveSources, tags: [world], x-palpanel-permission: read, responses: *readResponses}
...
  /save-sources/import:
    post:
      operationId: importSaveSource
      tags: [world]
      x-palpanel-permission: 'server:control'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: {$ref: '#/components/schemas/SaveSourceImportRequest'}
          application/json:
            schema: {$ref: '#/components/schemas/SaveImportCommitRequest'}
      responses:
        '200': *success
        '400': *error
        '401': *error
        '403': *error
        '409': *saveImportConflict
        '422': *error
        '500': *error
  /save-sources/{id}/activate:
    parameters: [*idParameter]
    post: {operationId: activateSaveSource, tags: [world], x-palpanel-permission: 'server:control', responses: *writeResponses}
  /save-sources/{id}/rebuild:
    parameters: [*idParameter]
    post: {operationId: rebuildSaveSource, tags: [world], x-palpanel-permission: 'server:control', responses: *writeResponses}
  /save-sources/{id}:
    parameters: [*idParameter]
    patch: {operationId: renameSaveSource, tags: [world], x-palpanel-permission: 'server:control', requestBody: *jsonRequest, responses: *writeResponses}
    delete: {operationId: deleteSaveSource, tags: [world], x-palpanel-permission: 'server:control', responses: *writeResponses}
...
  /breeding/jobs:
    post: {operationId: submitBreedingJob, tags: [world], x-palpanel-permission: read, requestBody: *jsonRequest, responses: *writeResponses}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: palpanel_session
      description: HttpOnly SameSite=Lax session cookie created by registration or login.
    developmentKey:
      type: http
      scheme: bearer
      bearerFormat: ppk_...
      description: 'Revocable development key supplied as Authorization: Bearer ppk_...'
  schemas:
...
    SaveSourceImportRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: Palworld world save archive in .zip, .tar, .tar.gz, or .tgz format.
        name:
          type: string
          description: Optional display name for the imported save source.
...
    SaveImportCommitRequest:
      type: object
      additionalProperties: false
      properties:
        inspection_id: {type: string, minLength: 1}
        candidate_id: {type: string, minLength: 1}
        name: {type: string}
        migration_source_id: {type: string, minLength: 1}
        steam_id: {type: string, pattern: '^[0-9]{15,20}$'}
        confirm: {type: boolean, const: true}
      oneOf:
        - required: [inspection_id]
        - required: [migration_source_id, steam_id, confirm]
...